+++ source/lue_cxx/include/lue/constant_size/time/omnipresent/space_domain.hpp
@@ -17,7 +17,7 @@ public:
 
                    SpaceDomain         (constant_size::SpaceDomain&& domain);
 
-                   SpaceDomain         (SpaceDomain const& other);
+                   SpaceDomain         (SpaceDomain const& other)=default;
 
                    SpaceDomain         (SpaceDomain&& other)=default;
 
@@ -30,7 +30,9 @@ public:
 };
 
 
-SpaceDomain        create_space_domain (hdf5::Group const& group);
+SpaceDomain        create_space_domain (hdf5::Group const& group,
+                                        SpaceDomain::Configuration const&
+                                            configuration);
 
 }  // namespace omnipresent
 }  // namespace time