+++ source/lue_cxx/include/lue/domain.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "lue/define.hpp"
 #include "lue/hdf5/group.hpp"
 // #include "lue/cxx_api/domain_configuration.h"
 // #include "lue/cxx_api/space_domain.h"
@@ -17,6 +18,36 @@ class Domain:
 
 public:
 
+    class Configuration
+    {
+
+    public:
+
+                   Configuration       (TimeDomainType const type);
+
+                   Configuration       (hdf5::Attributes const& attributes);
+
+                   Configuration       (Configuration const& other)=default;
+
+                   ~Configuration      ()=default;
+
+        Configuration& operator=       (Configuration const& other)=default;
+
+        TimeDomainType
+                   time_domain_type    () const;
+
+        void       save                (hdf5::Attributes& attributes) const;
+
+    private:
+
+        TimeDomainType
+                   _time_domain_type;
+
+        void       load                (hdf5::Attributes const& attributes);
+
+    };
+
+
                    Domain              (hdf5::Identifier const& location);
 
                    Domain              (hdf5::Group&& group);
@@ -31,6 +62,9 @@ public:
 
     Domain&        operator=           (Domain&& other)=default;
 
+    Configuration const&
+                   configuration       () const;
+
     // TimeDomain&    time_domain         () const;
 
     // SpaceDomain&   space_domain        () const;
@@ -39,6 +73,8 @@ public:
     //                configuration       () const;
 
 private:
+
+    Configuration  _configuration;
 
     // std::unique_ptr<TimeDomain> _time_domain;
 
@@ -49,8 +85,8 @@ private:
 };
 
 
-Domain             create_domain       (hdf5::Identifier const& location);
-                                        // DomainConfiguration const&
-                                        //     configuration);
+Domain             create_domain       (hdf5::Identifier const& location,
+                                        Domain::Configuration const&
+                                            configuration);
 
 } // namespace lue