+++ source/lue_cxx/include/lue/tag.hpp
@@ -11,10 +11,21 @@ static std::string const ids_tag = "lue_ids";
 static std::string const nr_items_tag = "lue_nr_items";
 static std::string const phenomena_tag = "lue_phenomena";
 static std::string const rank_tag = "lue_rank";
-static std::string const space_domain_tag = "lue_space_domain";
-static std::string const space_discretization_tag = "lue_space_discretization";
+static std::string const shape_per_item_type_tag = "lue_shape_per_item_type";
+
+static std::string const space_discretization_tag =
+    "lue_space_discretization";
+static std::string const space_domain_tag =
+    "lue_space_domain";
+static std::string const space_domain_item_type_tag =
+    "lue_space_domain_item_type";
+static std::string const space_domain_type_tag =
+    "lue_space_domain_type";
+
 static std::string const time_domain_tag = "lue_time_domain";
+static std::string const time_domain_type_tag = "lue_time_domain_type";
 static std::string const time_discretization_tag = "lue_time_discretization";
+
 static std::string const universes_tag = "lue_universes";
 static std::string const value_tag = "lue_value";