+++ source/lue_cxx/src/constant_size/time/omnipresent/domain.cpp
@@ -9,8 +9,8 @@ namespace omnipresent {
 Domain::Domain(
     hdf5::Identifier const& location)
 
-    : constant_size::Domain(location),
-      _space(id())
+    : constant_size::Domain(location) // ,
+      // _space(id())
 
 {
 }
@@ -19,33 +19,34 @@ Domain::Domain(
 Domain::Domain(
     constant_size::Domain&& domain)
 
-    : constant_size::Domain(std::forward<constant_size::Domain>(domain)),
-      _space(id())
+    : constant_size::Domain(std::forward<constant_size::Domain>(domain)) // ,
+      // _space(id())
 
 {
 }
 
 
-SpaceDomain const& Domain::space() const
-{
-    return _space;
-}
+// SpaceDomain const& Domain::space() const
+// {
+//     return _space;
+// }
+// 
+// 
+// SpaceDomain& Domain::space()
+// {
+//     return _space;
+// }
 
 
-SpaceDomain& Domain::space()
-{
-    return _space;
-}
-
-
-Domain create_domain(
-    hdf5::Identifier const& location)
-{
-    auto domain = constant_size::create_domain(location);
-    auto space = omnipresent::create_space_domain(domain);
-
-    return Domain(std::move(domain));
-}
+// Domain create_domain(
+//     hdf5::Identifier const& location,
+//     Domain::Configuration const& configuration)
+// {
+//     auto domain = constant_size::create_domain(location, configuration);
+//     auto space = omnipresent::create_space_domain(domain);
+// 
+//     return Domain(std::move(domain));
+// }
 
 }  // namespace omnipresent
 }  // namespace time